requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.11
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-bootstrap-modal-forms==2.2.0
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
folium==0.12.1.post1
idna==3.3
import-export==0.2.67.dev6
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.22.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejwt==2.0.1
sqlparse==0.4.2
tablib==3.1.0
typing==3.7.4.3
tzdata==2021.5
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
